Start first CS50 project, topic is bit manipulation of .bmp files in C
This is a first commit. Stego2.c takes at the command line the path to a 24 bit .bmp file and outputs the
file size in bytes, the RGB color values of the bottom left most pixel, and the width of the image in pixels.
This project might become a trivial steganographic program, or may become a trivial image manipulation
program.